returns 0 if AcqCommand is run

+++ src/main/java/frc/acq/AcqCommand.java
@@ -3,5 +3,5 @@ package frc.acq;
 public abstract class AcqCommand {
 
     //Will be called by subsystem to execute this command
-    public abstract double execute();
+    public abstract int execute();
 }
